requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
google-generativeai==0.3.0
pymysql==1.1.0
python-dotenv==1.0.0
pydantic==2.4.0
cryptography==41.0.7